requirements.txt
fastapi
requests
beautifulsoup4
lxml
psycopg2
python-dotenv
pyjwt